(feat) add babel transpiling for gulpfile
